Move Status to its own file.
